Feature/51 주문 내역이 있는 상품 삭제 시 예외 처리 (#80)
* fix: update success message for item deletion

ITEM_UPDATED -> ITEM_DELETED

* feat: add order item check before item deletion

* feat: add error message(ITEM_BE_DISABLED)

* feat: implement boolean deleted for items

# Conflicts:
#	backend/src/main/java/com/java/NBE4_5_1_8/domain/item/service/ItemService.java

+++ backend/src/main/java/com/java/NBE4_5_1_8/domain/item/service/ItemService.java
@@ -5,6 +5,7 @@ import com.java.NBE4_5_1_8.domain.item.entity.Category;
 import com.java.NBE4_5_1_8.domain.item.entity.Item;
 import com.java.NBE4_5_1_8.domain.item.repository.CategoryRepository;
 import com.java.NBE4_5_1_8.domain.item.repository.ItemRepository;
+import com.java.NBE4_5_1_8.domain.orderinfo.repository.OrderItemRepository;
 import com.java.NBE4_5_1_8.global.exception.ServiceException;
 import com.java.NBE4_5_1_8.global.message.ErrorMessage;
 import lombok.RequiredArgsConstructor;
@@ -26,6 +27,7 @@ import java.util.List;
 public class ItemService {
     private final ItemRepository itemRepository;
     private final CategoryRepository categoryRepository;
+    private final OrderItemRepository orderItemRepository;
 
     @Value("${file.upload-dir}")
     private String itemsDir;
@@ -81,7 +83,7 @@ public class ItemService {
     }
 
     public List<Item> getItemList() {
-        return itemRepository.findAll();
+        return itemRepository.findAllByDeletedFalse();
     }
 
     public Item getItemById(Long itemId) {
@@ -110,6 +112,14 @@ public class ItemService {
     }
 
     public void deleteItem(Item item) {
+
+        boolean hasOrderHistory = orderItemRepository.existsByItem(item);
+        if (hasOrderHistory) {
+            item.setDeleted(true);
+            itemRepository.save(item);
+            throw new ServiceException(HttpStatus.OK, ErrorMessage.ITEM_BE_DISABLED);
+        }
+
         itemRepository.delete(item);
     }
 }
